# black/parser: optimize deepcopying nodes