--- a/tests/data/pattern_matching_extras.py
+++ b/tests/data/pattern_matching_extras.py
@@ -1,3 +1,5 @@
+import match
+
 match something:
     case [a as b]:
         print(b)
@@ -7,3 +9,21 @@ match something:
         print(b)
     case Point(int() as x, int() as y):
         print(x, y)
+
+
+match = 1
+case: int = re.match(something)
+
+match re.match(case):
+    case type("match", match):
+        pass
+    case match:
+        pass
+
+
+def func(match: case, case: match) -> case:
+    match Something():
+        case another:
+            ...
+        case func(match, case):
+            ...